fix(sensor): address fisheye camera PR review from LuisPovedaCano
- Drop the EndPlay reset of the wide-angle sensor counter: it is a
  process-wide index used only to derive unique subobject names at
  construction, and resetting it on one sensor's EndPlay let a surviving
  sibling sensor collide with the next spawned sensor's indices. Also
  rename the counter from WIDE_ANGLE_LENS_SENSOR_COUNTER_COUNTER to
  WIDE_ANGLE_LENS_SENSOR_COUNTER (typo).
- Stop capturing 'this' in the wide-angle ENQUEUE_RENDER_COMMAND lambda.
  The render command can outlive the actor, so dereferencing 'this'
  inside the lambda was unsafe. Snapshot the render targets, capture
  render target, sampler and distort options before the dispatch, and
  guard the lambda body with a TWeakObjectPtr.
- Remove the duplicate 'Parameters->FaceSampler = Sampler' assignment in
  CameraModelUtil::AllocateDistortShaderParameters.
- Drop the unused static_cast<int32> in WideAngleLens GetImageWidth and
  GetImageHeight (both return uint32).

+++ Unreal/CarlaUnreal/Plugins/Carla/Source/Carla/Sensor/SceneCaptureSensor_WideAngleLens.h
@@ -43,12 +43,12 @@ public:
 
   uint32 GetImageWidth() const
   {
-    return static_cast<int32>(ImageWidth);
+    return ImageWidth;
   }
 
   uint32 GetImageHeight() const
   {
-    return static_cast<int32>(ImageHeight);
+    return ImageHeight;
   }
 
   UFUNCTION(BlueprintCallable, BlueprintPure)
